Revert CPU/GPU split; keep IT speedups in shared GPU runner
AINode strongly depends on the test-worker environment (model weight
cache, GPU presence), so a CPU-only profile isn't a clean isolation
boundary. Roll the metadata tests back into AINodeSharedClusterIT and
keep the existing AIClusterIT GPU runner.

What stays from the earlier commit:
- AINodeWrapper symlinks /data/ainode/models instead of copying it per
  fork (with a copy fallback when the FS rejects symlinks).
- AINodeConcurrentForecastIT loop count 100 -> 10 for a concurrency
  smoke check.
- prepareDataInTree/Table/Table2 and prepareDataForTableModel switched
  from per-row execute() to addBatch()/executeBatch() in chunks of 500.

What's reverted:
- Drop AINodeIT category, AINodeBasicIT, the AINodeIT Maven profile, and
  the cluster-it-ainode-cpu.yml workflow.
- Re-tag AINodeClusterConfigIT back to AIClusterIT.
- Move the 4 metadata tests (drop builtin model error / show builtin
  models, both dialects) back into AINodeSharedClusterIT.
- Drop the "tolerate missing weights cache" branch in AINodeWrapper —
  with no CPU-only runner the cache is always present.
