--- integration-test/src/main/java/org/apache/iotdb/itbase/category/AINodeIT.java
@@ -1,27 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-
-package org.apache.iotdb.itbase.category;
-
-/**
- * Marker for AINode integration tests that exercise only metadata / lifecycle paths and therefore
- * don't need a GPU. Tests tagged with this category can run on plain CPU runners; tests that drive
- * inference, forecasting, or device-binding still belong in {@link AIClusterIT}.
- */
-public interface AINodeIT {}
